features(stepped-forms): re-architect example pages and build support for stepped-forms
