+++ src/app/examples/consts/pokemon-list.const.ts
@@ -1,0 +1,1061 @@
+import { FormOption } from '@models/signal-form.model';
+
+export const pokemonList: FormOption<{ name: string; typings: string[] }>[] = [
+  {
+    label: 'Bulbasaur',
+    value: {
+      name: 'Bulbasaur',
+      typings: ['grass', 'poison'],
+    },
+  },
+  {
+    label: 'Ivysaur',
+    value: {
+      name: 'Ivysaur',
+      typings: ['grass', 'poison'],
+    },
+  },
+  {
+    label: 'Venusaur',
+    value: {
+      name: 'Venusaur',
+      typings: ['grass', 'poison'],
+    },
+  },
+  {
+    label: 'Charmander',
+    value: {
+      name: 'Charmander',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Charmeleon',
+    value: {
+      name: 'Charmeleon',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Charizard',
+    value: {
+      name: 'Charizard',
+      typings: ['fire', 'Flying'],
+    },
+  },
+  {
+    label: 'Squirtle',
+    value: {
+      name: 'Squirtle',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Wartortle',
+    value: {
+      name: 'Wartortle',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Blastoise',
+    value: {
+      name: 'Blastoise',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Caterpie',
+    value: {
+      name: 'Caterpie',
+      typings: ['Bug'],
+    },
+  },
+  {
+    label: 'Metapod',
+    value: {
+      name: 'Metapod',
+      typings: ['Bug'],
+    },
+  },
+  {
+    label: 'Butterfree',
+    value: {
+      name: 'Butterfree',
+      typings: ['Bug', 'Flying'],
+    },
+  },
+  {
+    label: 'Weedle',
+    value: {
+      name: 'Weedle',
+      typings: ['Bug', 'poison'],
+    },
+  },
+  {
+    label: 'Kakuna',
+    value: {
+      name: 'Kakuna',
+      typings: ['Bug', 'poison'],
+    },
+  },
+  {
+    label: 'Beedrill',
+    value: {
+      name: 'Beedrill',
+      typings: ['Bug', 'poison'],
+    },
+  },
+  {
+    label: 'Pidgey',
+    value: {
+      name: 'Pidgey',
+      typings: ['Normal', 'Flying'],
+    },
+  },
+  {
+    label: 'Pidgeotto',
+    value: {
+      name: 'Pidgeotto',
+      typings: ['Normal', 'Flying'],
+    },
+  },
+  {
+    label: 'Pidgeot',
+    value: {
+      name: 'Pidgeot',
+      typings: ['Normal', 'Flying'],
+    },
+  },
+  {
+    label: 'Rattata',
+    value: {
+      name: 'Rattata',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Raticate',
+    value: {
+      name: 'Raticate',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Spearow',
+    value: {
+      name: 'Spearow',
+      typings: ['Normal', 'Flying'],
+    },
+  },
+  {
+    label: 'Fearow',
+    value: {
+      name: 'Fearow',
+      typings: ['Normal', 'Flying'],
+    },
+  },
+  {
+    label: 'Ekans',
+    value: {
+      name: 'Ekans',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Arbok',
+    value: {
+      name: 'Arbok',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Pikachu',
+    value: {
+      name: 'Pikachu',
+      typings: ['Electric'],
+    },
+  },
+  {
+    label: 'Raichu',
+    value: {
+      name: 'Raichu',
+      typings: ['Electric'],
+    },
+  },
+  {
+    label: 'Sandshrew',
+    value: {
+      name: 'Sandshrew',
+      typings: ['Ground'],
+    },
+  },
+  {
+    label: 'Sandslash',
+    value: {
+      name: 'Sandslash',
+      typings: ['Ground'],
+    },
+  },
+  {
+    label: 'Nidoran♀',
+    value: {
+      name: 'Nidoran♀',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Nidorina',
+    value: {
+      name: 'Nidorina',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Nidoqueen',
+    value: {
+      name: 'Nidoqueen',
+      typings: ['poison', 'Ground'],
+    },
+  },
+  {
+    label: 'Nidoran♂',
+    value: {
+      name: 'Nidoran♂',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Nidorino',
+    value: {
+      name: 'Nidorino',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Nidoking',
+    value: {
+      name: 'Nidoking',
+      typings: ['poison', 'Ground'],
+    },
+  },
+  {
+    label: 'Clefairy',
+    value: {
+      name: 'Clefairy',
+      typings: ['Fairy'],
+    },
+  },
+  {
+    label: 'Clefable',
+    value: {
+      name: 'Clefable',
+      typings: ['Fairy'],
+    },
+  },
+  {
+    label: 'Vulpix',
+    value: {
+      name: 'Vulpix',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Ninetales',
+    value: {
+      name: 'Ninetales',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Jigglypuff',
+    value: {
+      name: 'Jigglypuff',
+      typings: ['Normal', 'Fairy'],
+    },
+  },
+  {
+    label: 'Wigglytuff',
+    value: {
+      name: 'Wigglytuff',
+      typings: ['Normal', 'Fairy'],
+    },
+  },
+  {
+    label: 'Zubat',
+    value: {
+      name: 'Zubat',
+      typings: ['poison', 'Flying'],
+    },
+  },
+  {
+    label: 'Golbat',
+    value: {
+      name: 'Golbat',
+      typings: ['poison', 'Flying'],
+    },
+  },
+  {
+    label: 'Oddish',
+    value: {
+      name: 'Oddish',
+      typings: ['grass', 'poison'],
+    },
+  },
+  {
+    label: 'Gloom',
+    value: {
+      name: 'Gloom',
+      typings: ['grass', 'poison'],
+    },
+  },
+  {
+    label: 'Vileplume',
+    value: {
+      name: 'Vileplume',
+      typings: ['grass', 'poison'],
+    },
+  },
+  {
+    label: 'Paras',
+    value: {
+      name: 'Paras',
+      typings: ['Bug', 'grass'],
+    },
+  },
+  {
+    label: 'Parasect',
+    value: {
+      name: 'Parasect',
+      typings: ['Bug', 'grass'],
+    },
+  },
+  {
+    label: 'Venonat',
+    value: {
+      name: 'Venonat',
+      typings: ['Bug', 'poison'],
+    },
+  },
+  {
+    label: 'Venomoth',
+    value: {
+      name: 'Venomoth',
+      typings: ['Bug', 'poison'],
+    },
+  },
+  {
+    label: 'Diglett',
+    value: {
+      name: 'Diglett',
+      typings: ['Ground'],
+    },
+  },
+  {
+    label: 'Dugtrio',
+    value: {
+      name: 'Dugtrio',
+      typings: ['Ground'],
+    },
+  },
+  {
+    label: 'Meowth',
+    value: {
+      name: 'Meowth',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Persian',
+    value: {
+      name: 'Persian',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Psyduck',
+    value: {
+      name: 'Psyduck',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Golduck',
+    value: {
+      name: 'Golduck',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Mankey',
+    value: {
+      name: 'Mankey',
+      typings: ['Fighting'],
+    },
+  },
+  {
+    label: 'Primeape',
+    value: {
+      name: 'Primeape',
+      typings: ['Fighting'],
+    },
+  },
+  {
+    label: 'Growlithe',
+    value: {
+      name: 'Growlithe',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Arcanine',
+    value: {
+      name: 'Arcanine',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Poliwag',
+    value: {
+      name: 'Poliwag',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Poliwhirl',
+    value: {
+      name: 'Poliwhirl',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Poliwrath',
+    value: {
+      name: 'Poliwrath',
+      typings: ['Water', 'Fighting'],
+    },
+  },
+  {
+    label: 'Abra',
+    value: {
+      name: 'Abra',
+      typings: ['Psychic'],
+    },
+  },
+  {
+    label: 'Kadabra',
+    value: {
+      name: 'Kadabra',
+      typings: ['Psychic'],
+    },
+  },
+  {
+    label: 'Alakazam',
+    value: {
+      name: 'Alakazam',
+      typings: ['Psychic'],
+    },
+  },
+  {
+    label: 'Machop',
+    value: {
+      name: 'Machop',
+      typings: ['Fighting'],
+    },
+  },
+  {
+    label: 'Machoke',
+    value: {
+      name: 'Machoke',
+      typings: ['Fighting'],
+    },
+  },
+  {
+    label: 'Machamp',
+    value: {
+      name: 'Machamp',
+      typings: ['Fighting'],
+    },
+  },
+  {
+    label: 'Bellsprout',
+    value: {
+      name: 'Bellsprout',
+      typings: ['grass', 'poison'],
+    },
+  },
+  {
+    label: 'Weepinbell',
+    value: {
+      name: 'Weepinbell',
+      typings: ['grass', 'poison'],
+    },
+  },
+  {
+    label: 'Victreebel',
+    value: {
+      name: 'Victreebel',
+      typings: ['grass', 'poison'],
+    },
+  },
+  {
+    label: 'Tentacool',
+    value: {
+      name: 'Tentacool',
+      typings: ['Water', 'poison'],
+    },
+  },
+  {
+    label: 'Tentacruel',
+    value: {
+      name: 'Tentacruel',
+      typings: ['Water', 'poison'],
+    },
+  },
+  {
+    label: 'Geodude',
+    value: {
+      name: 'Geodude',
+      typings: ['Rock', 'Ground'],
+    },
+  },
+  {
+    label: 'Graveler',
+    value: {
+      name: 'Graveler',
+      typings: ['Rock', 'Ground'],
+    },
+  },
+  {
+    label: 'Golem',
+    value: {
+      name: 'Golem',
+      typings: ['Rock', 'Ground'],
+    },
+  },
+  {
+    label: 'Ponyta',
+    value: {
+      name: 'Ponyta',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Rapidash',
+    value: {
+      name: 'Rapidash',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Slowpoke',
+    value: {
+      name: 'Slowpoke',
+      typings: ['Water', 'Psychic'],
+    },
+  },
+  {
+    label: 'Slowbro',
+    value: {
+      name: 'Slowbro',
+      typings: ['Water', 'Psychic'],
+    },
+  },
+  {
+    label: 'Magnemite',
+    value: {
+      name: 'Magnemite',
+      typings: ['Electric', 'Steel'],
+    },
+  },
+  {
+    label: 'Magneton',
+    value: {
+      name: 'Magneton',
+      typings: ['Electric', 'Steel'],
+    },
+  },
+  {
+    label: 'Farfetch’d',
+    value: {
+      name: 'Farfetch’d',
+      typings: ['Normal', 'Flying'],
+    },
+  },
+  {
+    label: 'Doduo',
+    value: {
+      name: 'Doduo',
+      typings: ['Normal', 'Flying'],
+    },
+  },
+  {
+    label: 'Dodrio',
+    value: {
+      name: 'Dodrio',
+      typings: ['Normal', 'Flying'],
+    },
+  },
+  {
+    label: 'Seel',
+    value: {
+      name: 'Seel',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Dewgong',
+    value: {
+      name: 'Dewgong',
+      typings: ['Water', 'Ice'],
+    },
+  },
+  {
+    label: 'Grimer',
+    value: {
+      name: 'Grimer',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Muk',
+    value: {
+      name: 'Muk',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Shellder',
+    value: {
+      name: 'Shellder',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Cloyster',
+    value: {
+      name: 'Cloyster',
+      typings: ['Water', 'Ice'],
+    },
+  },
+  {
+    label: 'Gastly',
+    value: {
+      name: 'Gastly',
+      typings: ['Ghost', 'poison'],
+    },
+  },
+  {
+    label: 'Haunter',
+    value: {
+      name: 'Haunter',
+      typings: ['Ghost', 'poison'],
+    },
+  },
+  {
+    label: 'Gengar',
+    value: {
+      name: 'Gengar',
+      typings: ['Ghost', 'poison'],
+    },
+  },
+  {
+    label: 'Onix',
+    value: {
+      name: 'Onix',
+      typings: ['Rock', 'Ground'],
+    },
+  },
+  {
+    label: 'Drowzee',
+    value: {
+      name: 'Drowzee',
+      typings: ['Psychic'],
+    },
+  },
+  {
+    label: 'Hypno',
+    value: {
+      name: 'Hypno',
+      typings: ['Psychic'],
+    },
+  },
+  {
+    label: 'Krabby',
+    value: {
+      name: 'Krabby',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Kingler',
+    value: {
+      name: 'Kingler',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Voltorb',
+    value: {
+      name: 'Voltorb',
+      typings: ['Electric'],
+    },
+  },
+  {
+    label: 'Electrode',
+    value: {
+      name: 'Electrode',
+      typings: ['Electric'],
+    },
+  },
+  {
+    label: 'Exeggcute',
+    value: {
+      name: 'Exeggcute',
+      typings: ['grass', 'Psychic'],
+    },
+  },
+  {
+    label: 'Exeggutor',
+    value: {
+      name: 'Exeggutor',
+      typings: ['grass', 'Psychic'],
+    },
+  },
+  {
+    label: 'Cubone',
+    value: {
+      name: 'Cubone',
+      typings: ['Ground'],
+    },
+  },
+  {
+    label: 'Marowak',
+    value: {
+      name: 'Marowak',
+      typings: ['Ground'],
+    },
+  },
+  {
+    label: 'Hitmonlee',
+    value: {
+      name: 'Hitmonlee',
+      typings: ['Fighting'],
+    },
+  },
+  {
+    label: 'Hitmonchan',
+    value: {
+      name: 'Hitmonchan',
+      typings: ['Fighting'],
+    },
+  },
+  {
+    label: 'Lickitung',
+    value: {
+      name: 'Lickitung',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Koffing',
+    value: {
+      name: 'Koffing',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Weezing',
+    value: {
+      name: 'Weezing',
+      typings: ['poison'],
+    },
+  },
+  {
+    label: 'Rhyhorn',
+    value: {
+      name: 'Rhyhorn',
+      typings: ['Ground', 'Rock'],
+    },
+  },
+  {
+    label: 'Rhydon',
+    value: {
+      name: 'Rhydon',
+      typings: ['Ground', 'Rock'],
+    },
+  },
+  {
+    label: 'Chansey',
+    value: {
+      name: 'Chansey',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Tangela',
+    value: {
+      name: 'Tangela',
+      typings: ['grass'],
+    },
+  },
+  {
+    label: 'Kangaskhan',
+    value: {
+      name: 'Kangaskhan',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Horsea',
+    value: {
+      name: 'Horsea',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Seadra',
+    value: {
+      name: 'Seadra',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Goldeen',
+    value: {
+      name: 'Goldeen',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Seaking',
+    value: {
+      name: 'Seaking',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Staryu',
+    value: {
+      name: 'Staryu',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Starmie',
+    value: {
+      name: 'Starmie',
+      typings: ['Water', 'Psychic'],
+    },
+  },
+  {
+    label: 'Mr. Mime',
+    value: {
+      name: 'Mr. Mime',
+      typings: ['Psychic', 'Fairy'],
+    },
+  },
+  {
+    label: 'Scyther',
+    value: {
+      name: 'Scyther',
+      typings: ['Bug', 'Flying'],
+    },
+  },
+  {
+    label: 'Jynx',
+    value: {
+      name: 'Jynx',
+      typings: ['Ice', 'Psychic'],
+    },
+  },
+  {
+    label: 'Electabuzz',
+    value: {
+      name: 'Electabuzz',
+      typings: ['Electric'],
+    },
+  },
+  {
+    label: 'Magmar',
+    value: {
+      name: 'Magmar',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Pinsir',
+    value: {
+      name: 'Pinsir',
+      typings: ['Bug'],
+    },
+  },
+  {
+    label: 'Tauros',
+    value: {
+      name: 'Tauros',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Magikarp',
+    value: {
+      name: 'Magikarp',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Gyarados',
+    value: {
+      name: 'Gyarados',
+      typings: ['Water', 'Flying'],
+    },
+  },
+  {
+    label: 'Lapras',
+    value: {
+      name: 'Lapras',
+      typings: ['Water', 'Ice'],
+    },
+  },
+  {
+    label: 'Ditto',
+    value: {
+      name: 'Ditto',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Eevee',
+    value: {
+      name: 'Eevee',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Vaporeon',
+    value: {
+      name: 'Vaporeon',
+      typings: ['Water'],
+    },
+  },
+  {
+    label: 'Jolteon',
+    value: {
+      name: 'Jolteon',
+      typings: ['Electric'],
+    },
+  },
+  {
+    label: 'Flareon',
+    value: {
+      name: 'Flareon',
+      typings: ['fire'],
+    },
+  },
+  {
+    label: 'Porygon',
+    value: {
+      name: 'Porygon',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Omanyte',
+    value: {
+      name: 'Omanyte',
+      typings: ['Rock', 'Water'],
+    },
+  },
+  {
+    label: 'Omastar',
+    value: {
+      name: 'Omastar',
+      typings: ['Rock', 'Water'],
+    },
+  },
+  {
+    label: 'Kabuto',
+    value: {
+      name: 'Kabuto',
+      typings: ['Rock', 'Water'],
+    },
+  },
+  {
+    label: 'Kabutops',
+    value: {
+      name: 'Kabutops',
+      typings: ['Rock', 'Water'],
+    },
+  },
+  {
+    label: 'Aerodactyl',
+    value: {
+      name: 'Aerodactyl',
+      typings: ['Rock', 'Flying'],
+    },
+  },
+  {
+    label: 'Snorlax',
+    value: {
+      name: 'Snorlax',
+      typings: ['Normal'],
+    },
+  },
+  {
+    label: 'Articuno',
+    value: {
+      name: 'Articuno',
+      typings: ['Ice', 'Flying'],
+    },
+  },
+  {
+    label: 'Zapdos',
+    value: {
+      name: 'Zapdos',
+      typings: ['Electric', 'Flying'],
+    },
+  },
+  {
+    label: 'Moltres',
+    value: {
+      name: 'Moltres',
+      typings: ['fire', 'Flying'],
+    },
+  },
+  {
+    label: 'Dratini',
+    value: {
+      name: 'Dratini',
+      typings: ['Dragon'],
+    },
+  },
+  {
+    label: 'Dragonair',
+    value: {
+      name: 'Dragonair',
+      typings: ['Dragon'],
+    },
+  },
+  {
+    label: 'Dragonite',
+    value: {
+      name: 'Dragonite',
+      typings: ['Dragon', 'Flying'],
+    },
+  },
+  {
+    label: 'Mewtwo',
+    value: {
+      name: 'Mewtwo',
+      typings: ['Psychic'],
+    },
+  },
+  {
+    label: 'Mew',
+    value: {
+      name: 'Mew',
+      typings: ['Psychic'],
+    },
+  },
+] as const;
