+++ src/app/examples/example-signal-form-rows/example-signal-form-rows.component.scss
@@ -1,0 +1,23 @@
+:host {
+  max-width: 800px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  gap: 8px;
+  text-align: left;
+}
+
+.example-content {
+  margin-top: 2rem;
+
+  &-header {
+    width: 100%;
+    margin: 0;
+    padding: 0;
+  }
+
+  &-content {
+    font-size: 0.85rem;
+  }
+}
