+++ src/app/examples/example-signal-form-stepper/example-signal-form-stepper.component.html
@@ -1,0 +1,1 @@
+<signal-form-stepper [form]="steppedForm" />
