+++ src/app/examples/example-signal-form/example-signal-form.component.html
@@ -1,0 +1,3 @@
+@if (!!form) {
+  <app-signal-form [form]="form" (onSave)="save($event)" />
+}
