+++ src/app/signal-forms/components/fields/form-checkbox-group-field/form-checkbox-group-field.component.ts
@@ -5,9 +5,9 @@ import {
   computed,
   input,
 } from '@angular/core';
-import { FormFieldType } from '../../../enums/form-field-type.enum';
-import { FormOption } from '../../../models/signal-form.model';
-import { BaseInputDirective } from '../../base/base-input/base-input.directive';
+import { BaseInputDirective } from '@base/base-input/base-input.directive';
+import { FormFieldType } from '@enums/form-field-type.enum';
+import { type FormOption } from '@models/signal-form.model';
 
 @Component({
   selector: 'signal-form-checkbox-group-field',
