+++ src/app/signal-forms/components/fields/form-chip-list-field/form-chip-list-field.component.ts
@@ -1,8 +1,8 @@
 import { Component, computed } from '@angular/core';
-import { FormFieldType } from '../../../enums/form-field-type.enum';
-import { ChipListFieldConfig } from '../../../models/signal-field-configs.model';
-import { FormOption } from '../../../models/signal-form.model';
-import { BaseInputDirective } from '../../base/base-input/base-input.directive';
+import { BaseInputDirective } from '@base/base-input/base-input.directive';
+import { FormFieldType } from '@enums/form-field-type.enum';
+import { type ChipListFieldConfig } from '@models/signal-field-configs.model';
+import { type FormOption } from '@models/signal-form.model';
 
 @Component({
   selector: 'signal-form-chip-list-field',
