+++ src/app/signal-forms/components/fields/form-number-field/form-number-field.component.html
@@ -5,6 +5,8 @@
     [attr.aria-describedby]="ariaDescribedBy()"
     [attr.aria-invalid]="error() ? 'true' : 'false'"
     [attr.name]="name()"
+    [required]="isRequired()"
+    [attr.aria-required]="isRequired()"
     [id]="name()"
     [placeholder]="config()?.placeholder"
     [disabled]="disabled()"
