+++ src/app/signal-forms/components/fields/form-password-field/form-password-field.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { FormFieldType } from '../../../enums/form-field-type.enum';
-import { type PasswordFieldConfig } from '../../../models/signal-field-configs.model';
-import { BaseInputDirective } from '../../base/base-input/base-input.directive';
+import { BaseInputDirective } from '@base/base-input/base-input.directive';
+import { FormFieldType } from '@enums/form-field-type.enum';
+import { type PasswordFieldConfig } from '@models/signal-field-configs.model';
 
 @Component({
   selector: 'signal-form-password-field',
