+++ src/app/signal-forms/components/fields/form-slider-field/form-slider-field.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { FormFieldType } from '../../../enums/form-field-type.enum';
-import { type SliderFieldConfig } from '../../../models/signal-field-configs.model';
-import { BaseInputDirective } from '../../base/base-input/base-input.directive';
+import { BaseInputDirective } from '@base/base-input/base-input.directive';
+import { FormFieldType } from '@enums/form-field-type.enum';
+import { type SliderFieldConfig } from '@models/signal-field-configs.model';
 
 @Component({
   selector: 'signal-form-slider-field',
