+++ src/app/signal-forms/components/renderers/form-field-error-summary/form-error-summary.component.scss
@@ -17,5 +17,6 @@
     background: transparent;
     color: white;
     outline: none;
+    cursor: pointer;
   }
 }
