+++ src/app/signal-forms/components/renderers/signal-form-save-button/signal-form-save-button.component.ts
@@ -4,7 +4,7 @@ import {
   input,
   output,
 } from '@angular/core';
-import { type SignalFormContainer } from '../../../models/signal-form.model';
+import { type SignalFormContainer } from '@models/signal-form.model';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
