+++ src/app/signal-forms/components/renderers/signal-form-stepper/signal-form-stepper-nav/signal-form-stepper-nav.component.scss
@@ -1,0 +1,60 @@
+.stepper-nav-ui {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 1rem;
+
+  .stepper-track {
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+
+    .stepper-node {
+      display: flex;
+      align-items: center;
+      gap: 1rem;
+
+      .step-circle {
+        width: 32px;
+        height: 32px;
+        border-radius: 50%;
+        border: 2px solid #ccc;
+        background-color: white;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        font-weight: bold;
+        transition: all 0.2s ease;
+        cursor: pointer;
+
+        &.active {
+          border-color: #4f46e5 !important;
+        }
+
+        &.completed {
+          background-color: #4ade80;
+          color: white;
+          border-color: #4ade80;
+        }
+
+        &.error {
+          border-color: #ef4444;
+          color: #ef4444;
+        }
+      }
+
+      .step-line {
+        width: 48px;
+        height: 2px;
+        background: #e5e7eb;
+
+        &.complete {
+          background: #4ade80;
+        }
+
+        &.error {
+          background: #ef4444;
+        }
+      }
+    }
+  }
+}
