+++ src/app/signal-forms/components/renderers/signal-form-stepper/signal-form-stepper.component.html
@@ -1,0 +1,48 @@
+@if (form(); as form) {
+  <form onsubmit="event.preventDefault();">
+    <signal-form-stepper-nav
+      [steps]="form.steps"
+      [(currentStep)]="form.currentStep"
+    />
+
+    <div class="stepper-body">
+      @for (step of form.steps; track $index) {
+        <div [hidden]="$index !== form.currentStep()">
+          @defer (on viewport; prefetch on idle) {
+            <signal-form-fields
+              [form]="step"
+              [fields]="step.fields"
+              #formFields
+            />
+          } @placeholder {
+            <div></div>
+          }
+        </div>
+      }
+    </div>
+
+    <signal-form-error-summary [form]="form" />
+
+    <div class="stepper-actions">
+      <button
+        class="form-button"
+        (click)="previous()"
+        [disabled]="form.currentStep() === 0"
+      >
+        Previous
+      </button>
+      @if (form.currentStep() < form.steps.length - 1) {
+        <button
+          class="form-button"
+          (click)="next()"
+          [disabled]="!form.isValidStep()"
+        >
+          Next
+        </button>
+      }
+      @if (form.currentStep() === form.steps.length - 1) {
+        <button class="form-button" (click)="save()">Submit</button>
+      }
+    </div>
+  </form>
+}
