+++ src/app/signal-forms/components/renderers/signal-form-stepper/signal-form-stepper.component.scss
@@ -1,0 +1,8 @@
+.stepper-actions {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.stepper-body {
+  margin-bottom: 1rem;
+}
