+++ src/app/signal-forms/components/renderers/signal-form/signal-form.component.ts
@@ -5,8 +5,8 @@ import {
   input,
   output,
 } from '@angular/core';
-import { type SignalFormContainer } from '../../../models/signal-form.model';
-import { SignalFormHostDirective } from '../../base/host-directive/signal-form-host.directive';
+import { SignalFormHostDirective } from '@base/host-directive/signal-form-host.directive';
+import { type SignalFormContainer } from '@models/signal-form.model';
 import { SignalFormErrorSummaryComponent } from '../form-field-error-summary/signal-form-error-summary.component';
 import { SignalFormFieldsComponent } from '../signal-form-fields/signal-form-fields.component';
 
@@ -18,11 +18,14 @@ import { SignalFormFieldsComponent } from '../signal-form-fields/signal-form-fie
   styleUrl: './signal-form.component.scss',
   templateUrl: './signal-form.component.html',
   hostDirectives: [SignalFormHostDirective],
+  host: {
+    '[class.signal-form-container]': 'true',
+  },
 })
 export class SignalFormComponent<TModel> {
   public form = input.required<SignalFormContainer<TModel>>();
-  submitButtonText = input<string>('Save');
-  onSave = output<TModel>();
+  public submitButtonText = input<string>('Save');
+  public onSave = output<TModel>();
 
   protected readonly disabled = computed(() => {
     const anyTouched = this.form().anyTouched();
@@ -31,7 +34,7 @@ export class SignalFormComponent<TModel> {
     return !(anyTouched && anyDirty);
   });
 
-  submitForm(): void {
+  protected submitForm(): void {
     this.form().save();
     this.onSave.emit(this.form().getValue());
   }
