+++ src/app/signal-forms/enums/form-field-type.enum.ts
@@ -1,21 +1,21 @@
 export enum FormFieldType {
-  TEXT = 'text',
-  PASSWORD = 'password',
-  TEXTAREA = 'textarea',
-  SELECT = 'select',
-  RADIO = 'radio',
   AUTOCOMPLETE = 'autocomplete',
-  NUMBER = 'number',
   CHECKBOX = 'checkbox',
   CHECKBOX_GROUP = 'checkbox-group',
-  DATETIME = 'datetime',
+  CHIPLIST = 'chiplist',
   COLOR = 'color',
-  SWITCH = 'switch',
-  SLIDER = 'slider',
+  DATETIME = 'datetime',
   FILE = 'file',
-  RATING = 'rating',
   MASKED = 'masked',
   MULTISELECT = 'multiselect',
-  CHIPLIST = 'chiplist',
+  NUMBER = 'number',
+  PASSWORD = 'password',
+  RADIO = 'radio',
+  RATING = 'rating',
   REPEATABLE_GROUP = 'repeatable-group',
+  SELECT = 'select',
+  SLIDER = 'slider',
+  SWITCH = 'switch',
+  TEXT = 'text',
+  TEXTAREA = 'textarea',
 }
