+++ src/styles/forms.scss
@@ -1,3 +1,11 @@
+.signal-form-container {
+  padding: 1rem;
+  border-radius: 4px;
+  display: block;
+  width: 600px;
+  background: white;
+}
+
 .form-control {
   display: flex;
   flex-direction: column;
